backend/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
bcrypt==3.2.0
bidict==0.22.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
databases==0.5.5
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.1
fastapi-mail==1.0.8
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-cloud-translate==3.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-engineio==4.3.1
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.5.2
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
wrapt==1.14.0
